requirements.txt
PyMySQL
flask-cors
flask_sqlalchemy
flask_script
flask_migrate
grequests
asyncio